Changes wip names to Craps and CrapsConsole
